crawler-scripts/requirements.txt
requests
beautifulsoup4
playwright
pymongo
python-dotenv
